requirements.txt
PyQt5>=5.15.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
anthropic>=0.89.0
openai>=1.0.0
httpx>=0.28.0
requests>=2.25.1
python-dotenv>=1.0.0
